chore: suppression class (#374)
