+++ includes/class-core.php
@@ -94,7 +94,8 @@ final class Core {
 		include_once NEWSPACK_ADS_ABSPATH . '/includes/integrations/class-ad-refresh-control.php';
 		include_once NEWSPACK_ADS_ABSPATH . '/includes/class-blocks.php';
 		include_once NEWSPACK_ADS_ABSPATH . '/includes/class-widget.php';
-		include_once NEWSPACK_ADS_ABSPATH . '/includes/suppress-ads.php';
+		include_once NEWSPACK_ADS_ABSPATH . '/includes/class-suppression.php';
+		include_once NEWSPACK_ADS_ABSPATH . '/includes/functions.php';
 	}
 
 	/**
